Reorganized test files
